agency entity -- package renaming
-change package name gtfsentity to gtfs
