+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/Agency.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.domain.entity.gtfsentity;
+package org.mobilitydata.gtfsvalidator.domain.entity.gtfs;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
