+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/GtfsEntity.java
@@ -1,0 +1,5 @@
+package org.mobilitydata.gtfsvalidator.domain.entity.gtfs;
+
+public abstract class GtfsEntity {
+
+}
