--- domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfsentity/GtfsEntity.java
@@ -1,5 +1,0 @@
-package org.mobilitydata.gtfsvalidator.domain.entity.gtfsentity;
-
-public abstract class GtfsEntity {
-
-}
